requirements.txt
openai
anthropic
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf